fix: test_utils is #[cfg(test)], and simpler
